login handled better
